+++ src/integrations/terminal/TerminalRegistry.ts
@@ -1,4 +1,5 @@
 import * as vscode from "vscode"
+import { arePathsEqual } from "../../utils/path"
 import { Terminal } from "./Terminal"
 import { TerminalProcess } from "./TerminalProcess"
 
@@ -173,5 +174,40 @@ export class TerminalRegistry {
 	static cleanup() {
 		this.disposables.forEach((disposable) => disposable.dispose())
 		this.disposables = []
+	}
+
+	/**
+	 * Gets an existing terminal or creates a new one for the given working directory
+	 * @param cwd The working directory path
+	 * @returns A Terminal instance
+	 */
+	static async getOrCreateTerminal(cwd: string): Promise<Terminal> {
+		const terminals = this.getAllTerminals()
+
+		// Find available terminal from our pool first (created for this task)
+		const matchingTerminal = terminals.find((t) => {
+			if (t.busy) {
+				return false
+			}
+			const terminalCwd = t.terminal.shellIntegration?.cwd // one of cline's commands could have changed the cwd of the terminal
+			if (!terminalCwd) {
+				return false
+			}
+			return arePathsEqual(vscode.Uri.file(cwd).fsPath, terminalCwd.fsPath)
+		})
+		if (matchingTerminal) {
+			return matchingTerminal
+		}
+
+		// If no matching terminal exists, try to find any non-busy terminal
+		const availableTerminal = terminals.find((t) => !t.busy)
+		if (availableTerminal) {
+			// Navigate back to the desired directory
+			await availableTerminal.runCommand(`cd "${cwd}"`)
+			return availableTerminal
+		}
+
+		// If all terminals are busy, create a new one
+		return this.createTerminal(cwd)
 	}
 }
